Remove unused functionality around search, download, and indexing
These were planned features that were started but never finished. The
idea was to captured the content of starred entries to the local
filesystem and make available for search via elasticsearch.

Search may be implemented in the future, but it probably will be
implemented in a different way and most likely won't use
elasticsearch.