Refactor insist() to work more like emit()
It was awkward having insist take event arguments as an array. Now its
signature takes better advantage of the arguments variable, and can
act like more a stand-in for emit the way it was intended. Failure is
also now conveyed back to the caller via the insist:failure event, and
insist is generally simpler.