Add a handler for applying filters to entries
The filter:apply event is called by the entry:store handler just after
a successful. It receives the entry and sets a user-specific score by
appling the weights of all available filters for the entry feed. The
ids of the filters that were matched are also stored, in the
userEntryFilters table.