Use createObjectUrl for OPML instead of FileSaver
FileSaver was used to get the browser to download an OPML export as a
standalone file. The export request happens inside the application
instead of as a standalone request in order to populate the
authentication header, which is why the export link doesn't just link
to the corresponding endpoint directly.

createObjectUrl accomplishes the same ends, but without requiring
another third-party library dependency. Safari still doesn't trigger a
download like Chrome does, but this was happening anyway with FileSaver.