Remove fetching via Google Feed API
The API is no longer being offered, and has started returning error
response which cause the fetcher to crash. Remove this form of
fetching as a temporary measure rather than back port the newer
solution from dispatcher. This will limit the application drastically
but will also motivate further work on the dispatcher.